requirements.txt
streamlit
pandas
openpyxl
google-generativeai
python-dotenv
pandas